+++ go/indexer/indexer.go
@@ -9,6 +9,10 @@ import (
 	"strconv"
 	"time"
 
+	"github.com/ethereum-optimism/optimism/go/indexer/services"
+
+	l2rpc "github.com/ethereum-optimism/optimism/l2geth/rpc"
+
 	"github.com/ethereum-optimism/optimism/go/indexer/metrics"
 	"github.com/ethereum-optimism/optimism/go/indexer/server"
 	"github.com/rs/cors"
@@ -81,8 +85,10 @@ type Indexer struct {
 
 	l1IndexingService *l1.Service
 	l2IndexingService *l2.Service
+	airdropService    *services.Airdrop
 
-	router *mux.Router
+	router  *mux.Router
+	metrics *metrics.Metrics
 }
 
 // NewIndexer initializes the Indexer, gathering any resources
@@ -128,7 +134,7 @@ func NewIndexer(cfg Config, gitVersion string) (*Indexer, error) {
 		return nil, err
 	}
 
-	l2Client, err := dialL2EthClientWithTimeout(ctx, cfg.L2EthRpc)
+	l2Client, l2RPC, err := dialL2EthClientWithTimeout(ctx, cfg.L2EthRpc)
 	if err != nil {
 		return nil, err
 	}
@@ -180,6 +186,7 @@ func NewIndexer(cfg Config, gitVersion string) (*Indexer, error) {
 	l2IndexingService, err := l2.NewService(l2.ServiceConfig{
 		Context:            ctx,
 		Metrics:            m,
+		L2RPC:              l2RPC,
 		L2Client:           l2Client,
 		DB:                 db,
 		ConfDepth:          cfg.ConfDepth,
@@ -198,7 +205,9 @@ func NewIndexer(cfg Config, gitVersion string) (*Indexer, error) {
 		l2Client:          l2Client,
 		l1IndexingService: l1IndexingService,
 		l2IndexingService: l2IndexingService,
+		airdropService:    services.NewAirdrop(db, m),
 		router:            mux.NewRouter(),
+		metrics:           m,
 	}, nil
 }
 
@@ -213,6 +222,7 @@ func (b *Indexer) Serve() error {
 	b.router.HandleFunc("/v1/deposits/0x{address:[a-fA-F0-9]{40}}", b.l1IndexingService.GetDeposits).Methods("GET")
 	b.router.HandleFunc("/v1/withdrawal/0x{hash:[a-fA-F0-9]{64}}", b.l2IndexingService.GetWithdrawalBatch).Methods("GET")
 	b.router.HandleFunc("/v1/withdrawals/0x{address:[a-fA-F0-9]{40}}", b.l2IndexingService.GetWithdrawals).Methods("GET")
+	b.router.HandleFunc("/v1/airdrops/0x{address:[a-fA-F0-9]{40}}", b.airdropService.GetAirdrop)
 	b.router.HandleFunc("/healthz", func(w http.ResponseWriter, r *http.Request) {
 		w.WriteHeader(200)
 		_, err := w.Write([]byte("OK"))
@@ -221,7 +231,7 @@ func (b *Indexer) Serve() error {
 		}
 	})
 
-	middleware := server.LoggingMiddleware(log.New("service", "server"))
+	middleware := server.LoggingMiddleware(b.metrics, log.New("service", "server"))
 
 	port := strconv.FormatUint(b.cfg.RESTPort, 10)
 	addr := fmt.Sprintf("%s:%s", b.cfg.RESTHostname, port)
@@ -277,12 +287,17 @@ func dialL1EthClientWithTimeout(ctx context.Context, url string) (
 // provided URL. If the dial doesn't complete within defaultDialTimeout seconds,
 // this method will return an error.
 func dialL2EthClientWithTimeout(ctx context.Context, url string) (
-	*l2ethclient.Client, error) {
+	*l2ethclient.Client, *l2rpc.Client, error) {
 
 	ctxt, cancel := context.WithTimeout(ctx, defaultDialTimeout)
 	defer cancel()
 
-	return l2ethclient.DialContext(ctxt, url)
+	rpc, err := l2rpc.DialContext(ctxt, url)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	return l2ethclient.NewClient(rpc), rpc, nil
 }
 
 // traceRateToFloat64 converts a time.Duration into a valid float64 for the