+++ l2geth/rollup/metrics.go
@@ -1,7 +1,0 @@
-package rollup
-
-import "github.com/ethereum-optimism/optimism/l2geth/metrics"
-
-var (
-	pubTxDropCounter = metrics.NewRegisteredCounter("rollup/pub/txdrops", nil)
-)