+++ go/indexer/CHANGELOG.md
@@ -1,6 +1,30 @@
 # @eth-optimism/indexer
 
+## 0.1.0
+
+### Minor Changes
+
+- 79be3e80: Add airdrops API
+
+### Patch Changes
+
+- d7bb9625: fix context reuse
+
+## 0.0.4
+
+### Patch Changes
+
+- 1df934a1: Don't spam the backend
+- 94876e28: fix (indexer): update l2 bridge addresses
+
+## 0.0.3
+
+### Patch Changes
+
+- 160f4c3d: Update docker image to use golang 1.18.0
+
 ## 0.0.2
+
 ### Patch Changes
 
 - 0e40dcb6: Indexer: initial release