+++ go/l2geth-exporter/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @eth-optimism/l2geth-exporter
 
+## 0.0.5
+
+### Patch Changes
+
+- 373d0e6f: Added SCC collection
+
+## 0.0.4
+
+### Patch Changes
+
+- 160f4c3d: Update docker image to use golang 1.18.0
+
 ## 0.0.3
 
 ### Patch Changes