+++ go/teleportr/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @eth-optimism/teleportr
 
+## 0.0.4
+
+### Patch Changes
+
+- 7a320e22: Use L2 gas price in driver
+
+## 0.0.3
+
+### Patch Changes
+
+- 160f4c3d: Update docker image to use golang 1.18.0
+
 ## 0.0.2
 
 ### Patch Changes