+++ l2geth/CHANGELOG.md
@@ -1,5 +1,30 @@
 # Changelog
 
+## 0.5.18
+
+### Patch Changes
+
+- 935a98e6: rollup: fix log.Crit usage
+- 81f09f16: l2geth: Record rollup transaction metrics
+
+## 0.5.17
+
+### Patch Changes
+
+- 13524da4: Style fix in the sync service
+- 160f4c3d: Update docker image to use golang 1.18.0
+- 1a28ba5f: Skip account cmd tests
+- 45582fcc: Skip unused tests in l2geth
+- 0c4d4e08: l2geth: Revert transaction pubsub feature
+
+## 0.5.16
+
+### Patch Changes
+
+- a01a2eb1: Skip TestWSAttachWelcome
+- 23ad6068: Skip some geth console tests that flake in CI
+- 6926b293: Adds a flag for changing the genesis fetch timeout
+
 ## 0.5.15
 
 ### Patch Changes