+++ packages/contracts/CHANGELOG.md
@@ -1,5 +1,29 @@
 # Changelog
 
+## 0.5.21
+
+### Patch Changes
+
+- 5818decb: Remove l2 gas price hardhat task
+
+## 0.5.20
+
+### Patch Changes
+
+- d040a8d9: Deleted update and helper functions/tests from Lib_MerkleTrie.sol and Lib_SecureMerkleTrie.sol
+- b57014d1: Update to typescript@4.6.2
+- Updated dependencies [b57014d1]
+  - @eth-optimism/core-utils@0.8.3
+
+## 0.5.19
+
+### Patch Changes
+
+- c1957126: Update Dockerfile to use Alpine
+- d9a51154: Bump to hardhat@2.9.1
+- Updated dependencies [c1957126]
+  - @eth-optimism/core-utils@0.8.2
+
 ## 0.5.18
 
 ### Patch Changes