+++ packages/core-utils/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @eth-optimism/core-utils
 
+## 0.8.3
+
+### Patch Changes
+
+- b57014d1: Update to typescript@4.6.2
+
+## 0.8.2
+
+### Patch Changes
+
+- c1957126: Update Dockerfile to use Alpine
+
 ## 0.8.1
 
 ### Patch Changes